// src/types/enums.ts
export enum ProjectType {
  redesign = "Редизайн",
  automatization = "Автоматизация",
  development = "Разработка",
  promotion = "Реклама",
  support = "Техподдержка",
  integration = "Интеграция",
}

export enum ClientType {
  law = "Юридическое",
  phys = "Физическое",
}

export enum ProjectStages {
  requirements = "Сбор требований",
  design = "Проектирование",
  development = "Разработка",
  testing = "Тестирование",
  ended = "Завершён",
  rejected = "Отклонён",
}

export enum ProjectStatuses {
  work = "В процессе",
  ended = "Завершён",
  rejected = "Отклонён",
}